+++ MechJeb.cs
@@ -39,7 +39,7 @@ namespace KRPC.MechJeb {
 
 			APIReady = false;
 			modules = new object[5];
-			windows = new object[3];
+			windows = new object[4];
 			controllers = new object[7];
 
 			Instance = FlightGlobals.ActiveVessel.GetMasterMechJeb();
@@ -96,6 +96,9 @@ namespace KRPC.MechJeb {
 
 		[KRPCProperty]
 		public static SmartRCS SmartRCS => GetComputerModule<SmartRCS>(windows, 1);
+
+		[KRPCProperty]
+		public static Translatron Translatron => GetComputerModule<Translatron>(windows, 3);
 
 		// CONTROLLERS